enhancement: move response structs to common types
